+++ src/main/java/roomescape/application/AuthService.java
@@ -4,9 +4,9 @@ package roomescape.application;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Service;
-import roomescape.application.dto.MemberResponse;
-import roomescape.application.dto.TokenRequest;
-import roomescape.application.dto.TokenResponse;
+import roomescape.dto.MemberResponse;
+import roomescape.dto.TokenRequest;
+import roomescape.dto.TokenResponse;
 import roomescape.domain.member.Email;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRepository;
