+++ src/main/java/roomescape/application/MemberService.java
@@ -2,8 +2,8 @@ package roomescape.application;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.application.dto.MemberResponse;
-import roomescape.application.dto.MemberSignUpRequest;
+import roomescape.dto.MemberResponse;
+import roomescape.dto.MemberSignUpRequest;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRepository;
