+++ src/main/java/roomescape/application/ReservationService.java
@@ -4,11 +4,11 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.application.dto.LoginMember;
-import roomescape.application.dto.MyReservationResponse;
-import roomescape.application.dto.ReservationCriteria;
-import roomescape.application.dto.ReservationRequest;
-import roomescape.application.dto.ReservationResponse;
+import roomescape.dto.LoginMember;
+import roomescape.dto.MyReservationResponse;
+import roomescape.dto.ReservationCriteria;
+import roomescape.dto.ReservationRequest;
+import roomescape.dto.ReservationResponse;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservation.ReservationFactory;
 import roomescape.domain.reservation.ReservationRepository;
