+++ src/main/java/roomescape/application/ThemeService.java
@@ -3,8 +3,8 @@ package roomescape.application;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.application.dto.ThemeRequest;
-import roomescape.application.dto.ThemeResponse;
+import roomescape.dto.ThemeRequest;
+import roomescape.dto.ThemeResponse;
 import roomescape.domain.theme.PopularThemeFinder;
 import roomescape.domain.theme.Theme;
 import roomescape.domain.theme.ThemeRepository;
