+++ src/main/java/roomescape/dto/AvailableTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.application.dto;
+package roomescape.dto;
 
 import java.time.LocalTime;
 import roomescape.domain.dto.AvailableTimeDto;
