+++ src/main/java/roomescape/dto/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.application.dto;
+package roomescape.dto;
 
 public record LoginMember(Long id, String name, String email, String role) {
 }
