+++ src/main/java/roomescape/dto/MyReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.application.dto;
+package roomescape.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
