+++ src/main/java/roomescape/dto/ReservationCriteria.java
@@ -1,4 +1,4 @@
-package roomescape.application.dto;
+package roomescape.dto;
 
 import java.time.LocalDate;
 import org.springframework.format.annotation.DateTimeFormat;
