+++ src/main/java/roomescape/dto/ThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.application.dto;
+package roomescape.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import roomescape.domain.theme.Theme;
