+++ src/main/java/roomescape/dto/ThemeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.application.dto;
+package roomescape.dto;
 
 import roomescape.domain.theme.Theme;
