+++ src/test/java/roomescape/domain/reservation/ReservationFactoryTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import roomescape.application.ServiceTest;
-import roomescape.application.dto.ReservationRequest;
+import roomescape.dto.ReservationRequest;
 import roomescape.exception.RoomescapeErrorCode;
 import roomescape.exception.RoomescapeException;
