+++ src/test/java/roomescape/presentation/acceptance/ThemeAcceptanceTest.java
@@ -8,8 +8,8 @@ import io.restassured.http.ContentType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import roomescape.application.dto.ThemeRequest;
-import roomescape.application.dto.ThemeResponse;
+import roomescape.dto.ThemeRequest;
+import roomescape.dto.ThemeResponse;
 import roomescape.domain.theme.ThemeRepository;
 
 class ThemeAcceptanceTest extends AcceptanceTest {
